requirements.txt
Cython
matplotlib==2.1.0
tensorflow==1.10
numpy==1.15
inflect==0.2.5
librosa==0.6.0
scipy==1.0.0
tensorboardX==1.1
Unidecode==1.0.22
pillow
torch==1.3.1
fastdtw
dtw-python
bokeh
pandas
git+https://github.com/NVIDIA/dllogger
pathlib
numba==0.46.0
